+++ returns/io.py
@@ -312,7 +312,7 @@ class IOResult(
         return self.from_result(self._inner_value.map(function))
 
     def bind(
-        self: 'IOResult[_ValueType, _ErrorType]',
+        self,
         function: Callable[
             [_ValueType],
             'IOResult[_NewValueType, _ErrorType]',
@@ -335,6 +335,31 @@ class IOResult(
 
         """
         raise NotImplementedError
+
+    def unify(
+        self,
+        function: Callable[
+            [_ValueType],
+            'IOResult[_NewValueType, _NewErrorType]',
+        ],
+    ) -> 'IOResult[_NewValueType, Union[_ErrorType, _NewErrorType]]':
+        """
+        Composes successful container with a function that returns a container.
+
+        .. code:: python
+
+          >>> from returns.io import IOResult, IOFailure, IOSuccess
+          >>> def bindable(string: str) -> IOResult[str, str]:
+          ...      if len(string) > 1:
+          ...          return IOSuccess(string + 'b')
+          ...      return IOFailure(string + 'c')
+          ...
+          >>> assert IOSuccess('aa').unify(bindable) == IOSuccess('aab')
+          >>> assert IOSuccess('a').unify(bindable) == IOFailure('ac')
+          >>> assert IOFailure('a').unify(bindable) == IOFailure('a')
+
+        """
+        return self.bind(function)  # type: ignore
 
     def bind_result(
         self,