+++ returns/pointfree.py
@@ -1,2 +1,3 @@
 from returns._generated.pointfree.bind import _bind as bind  # noqa: F401
 from returns._generated.pointfree.rescue import _rescue as rescue  # noqa: F401
+from returns._generated.pointfree.unify import _unify as unify  # noqa: F401